Changed to keep actual data in the ShParam structure
Before, it kept pointers to data.  This was cumbersome to use in actual code.

This version keeps a copy of the data.  So instead of using *(shParam.Argc), we now use  shParam.Argc.

Downside:  moving the Argv vector<String> consumes heap memory and time.  However, we are relying on data localization and reasonable STL efficiency to prevent memory leaks.