Update to work with Mega2560, change from STL to Array
Updated to work with Mega2560.  Primarily, change user command parameter to a Tmsh_paramP (pointer) since the struct doesn't fit in an AVR message.
STL:  Does not work with the current IDE release.  Changed from STL to Array.  Note this was done with a #define flag so a switchback should be simple. Future note:  Called it ARDUINOSSH, should change to ARDUINOSTL.
Cosmetics: eliminated unnecessary #include statements.